backend/requirements.txt
flask==2.0.1
python-chess>=1.9.3
sqlalchemy==1.4.23
numpy>=1.21.2
scipy>=1.7.1
gunicorn==20.1.0
python-dotenv==0.19.0